Add files via upload
Team Members: 1. Vishal Modi (104658093) 2. Chirag Patel (104699582) 3. Unnitkumar Metaliya (104702352) Technologies Used: Python: For scrapping the website. We have used a library named BeautifulSoup. This is the library designed particularly for scrapping purposes. PHP: For displaying final results from the database to the website. BootStarp: For styling the website. Excel: For database. Idea & Purpose: We were supposed to take any website and scrap the data from it and store it in a database. After that, we were supposed to display that database in the form of website. We have taken an eCommerce website named newegg.com for scrapping purposes. It is an eCommerce website for computer related parts and devices. We have taken major items which gets used in our day to day life. The goal of our final website was to show only best items which are available on newegg.com website. It will display only best items which are available at the moment depending on the item price and shipping details. How to run? You will have to run all the *scrapper.py files from command line. It will scrap the items from the site and will make a separate database files with its specific given name. After that, you will have to run display.php file from the host. It will then display the stuff from the datasets which got created after running .py files.